Split LowLevel directory
I have split LowLevel directory because the importing of big files
was not what I like and some of the types was logical to split in to separ
modules for future use (Error exceptions for example).

Current state is not definitely final:
 - Test needs to be moved according to new type structure.
 - Function readBody is in Line module which doesn't make sense...
 - Line name of the module is probably not the best choice either.

It the end this split is experiment if the new structure will work better than
the old one ;).